+++ r/Rduckhts/inst/duckhts_extension/duckhts.c
@@ -65,6 +65,9 @@ extern void register_kmer_udf_functions(duckdb_connection connection);
 extern void register_read_tabix_function(duckdb_connection connection);
 extern void register_read_gtf_function(duckdb_connection connection);
 extern void register_read_gff_function(duckdb_connection connection);
+/* genbank_reader.c */
+extern void register_read_genbank_function(duckdb_connection connection);
+extern void register_genbank_to_fasta_function(duckdb_connection connection);
 /* hts_meta_reader.c */
 extern void register_read_hts_header_function(duckdb_connection connection);
 extern void register_read_hts_index_function(duckdb_connection connection);
@@ -266,6 +269,8 @@ DUCKDB_EXTENSION_ENTRYPOINT(duckdb_connection connection,
     register_read_tabix_function(connection);
     register_read_gtf_function(connection);
     register_read_gff_function(connection);
+    register_read_genbank_function(connection);
+    register_genbank_to_fasta_function(connection);
     register_read_hts_header_function(connection);
     register_read_hts_index_function(connection);
     register_read_hts_index_spans_function(connection);